+++ poetry/console/commands/init.py
@@ -48,7 +48,7 @@ The <info>init</info> command creates a basic <comment>pyproject.toml</> file in
         self.line(
             [
                 "",
-                "This command will guide you through creating your <info>poetry.toml</> config.",
+                "This command will guide you through creating your <info>pyproject.toml</> config.",
                 "",
             ]
         )
@@ -296,7 +296,8 @@ The <info>init</info> command creates a basic <comment>pyproject.toml</> file in
     def _validate_license(self, license):
         from poetry.spdx import license_by_id
 
-        license_by_id(license)
+        if license:
+            license_by_id(license)
 
         return license