+++ poetry/utils/helpers.py
@@ -95,3 +95,12 @@ def try_load_dotenv():
     env_file_path = workdir_path / dotenv_file_name
     if env_file_path.is_file():
         load_dotenv(env_file_path)
+
+
+def get_http_basic_auth(repository_name):  # type: (str) -> tuple
+    config = Config.create("auth.toml")
+    repo_auth = config.setting("http-basic.{}".format(repository_name))
+    if repo_auth:
+        return repo_auth["username"], repo_auth["password"]
+    return None
+