+++ CHANGELOG.md
@@ -1,6 +1,40 @@
 # Change Log
 
-## [Unreleased]
+## [0.11.3] - 2018-07-26
+
+### Changed
+
+- Poetry now only uses [TOML Kit](https://github.com/sdispater/tomlkit) for TOML files manipulation.
+- Improved dependency resolution debug information.
+
+### Fixed
+
+- Fixed missing dependency information for some packages.
+- Fixed handling of single versions when packaging.
+- Fixed dependency information retrieval from `.zip` and `.bz2` archives.
+- Fixed searching for and installing packages from private repositories with authentication. (Thanks to [@MarcDufresne](https://github.com/MarcDufresne))
+- Fixed a potential error when checking the `pyproject.toml` validity. (Thanks to [@ojii](https://github.com/ojii))
+- Fixed the lock file not tracking the `extras` information from `pyproject.toml`. (Thanks to [@cauebs](https://github.com/cauebs))
+- Fixed missing trailing slash in the Simple API urls for private repositories. (Thanks to [@bradsbrown](https://github.com/bradsbrown))
+
+
+## [0.11.2] - 2018-07-03
+
+### Fixed
+
+- Fixed missing dependencies when resolving in some cases.
+- Fixed path dependencies not working in `dev-dependencies`.
+- Fixed license validation in `init`. (Thanks to [@cauebs](https://github.com/cauebs))
+
+
+## [0.11.1] - 2018-06-29
+
+### Fixed
+
+- Fixed an error when locking dependencies on Python 2.7.
+
+
+## [0.11.0] - 2018-06-28
 
 ### Added
 
@@ -412,7 +446,11 @@ Initial release
 
 
 
-[Unreleased]: https://github.com/sdispater/poetry/compare/0.10.3...master
+[Unreleased]: https://github.com/sdispater/poetry/compare/0.11.3...master
+[0.11.3]: https://github.com/sdispater/poetry/releases/tag/0.11.3
+[0.11.2]: https://github.com/sdispater/poetry/releases/tag/0.11.2
+[0.11.1]: https://github.com/sdispater/poetry/releases/tag/0.11.1
+[0.11.0]: https://github.com/sdispater/poetry/releases/tag/0.11.0
 [0.10.3]: https://github.com/sdispater/poetry/releases/tag/0.10.3
 [0.10.2]: https://github.com/sdispater/poetry/releases/tag/0.10.2
 [0.10.1]: https://github.com/sdispater/poetry/releases/tag/0.10.1